print stack trace after each syscall
This patch prints the stack trace of the traced process after
each system call when using -k flag. It uses libunwind to
unwind the stack and to obtain the function name pointed by
the IP.

Tested on Ubuntu 12.04 64 with the distribution version of
libunwind (0.99-0.3ubuntu1) and on CentOS 6.3 with libunwind
form the source code.

This code was originally taken from strace-plus of Philip J. Guo.

 v3:
   adjusted several datatype to use unsigned (Denys)
   fix sybol_name leak (Denys)
   workaround for libunwind bug with libpthread (unw_step does
        not return 1 when at the end of the stack)
   fix several problem in in autoconf script (Dmitry)
   use PATH_MAX for DSO file path buffer

 v2:
   fixed several identation issues
   removed a strlen in alloc_mmap_cache (Denys)
   reload mmap_cache after execv
   fixed a bug in libunwind_backtrace binary search (Masatake)
   check return value of _UPT_create
   use die_out_of_memory after Xalloc
   use exiting instead of !entering
   created a new separate unwind.c file
   removed include from defs.h
   added --with-libunwind to autoconf script
   cleaned up defs.h from several inclusions