unwind: handle invalid pid
In unwind part of tcb finalization, tcp->s_ent
was accessed to decide whether pending stacktrace in tcp->queue
should be printed or not.

However, s_ent can be NULL if invalid pid is given to strace
as an trace target. Here invalid pid means that the pid is
not associated with any process.

This patch handles such case.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>
Signed-off-by: Luca Clementi <luca.clementi@gmail.com>